credentials: profil have now higer priority to env
Until now the AK/SK in env always had priority to the one in profil.
(in config.json)
It was sometime confusing as it would ignore profil in oparameter if
env was set.

now:
if no profil are enter, env still have an higher priority to default profil.
otherwise profil have higher priority to env.
It is still posible to use partially complet profil, and complet
the profils informations with env
(examples have the region in config.json and AK/SK in env)

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>